requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
cffi==1.15.1
click==8.1.3
cryptography==38.0.4
Django==4.1.3
django-redis==5.2.0
djangorestframework==3.14.0
numpy==1.23.5
packaging==21.3
pandas==1.5.2
pycparser==2.21
pyparsing==3.0.9
pytdx==1.72
python-dateutil==2.8.2
pytz==2022.6
redis==4.3.5
six==1.16.0
sqlparse==0.4.3
